// sdk/packages/simplex/src/proto/mpcvault/platform/v1/api.ts
export interface ExecuteSigningRequestsResponse {
  error:
    | Error
    | undefined;
  /**
   * symbol_value:"3Eq21vXNB5s86c62bVuUfTerseghjkuUqRPBmGRJhyTA"
   * comment:tx_hash is the hash of the signed transaction, only set if status is STATUS_SUCCEEDED.This does not mean that the transaction it self is successful. It only means that the signing request has been signed.
   */
  txHash: string;
  /** signatures is the signatures of the raw message, only set if status is STATUS_SUCCEEDED. */
  signatures:
    | SignatureContainer
    | undefined;
  /** signed_transaction is the signed transaction for broadcast, only set if status is STATUS_SUCCEEDED. */
  signedTransaction: string;
}

export interface SignatureContainer {
  signatures: SignatureContainer_SignResponse[];
}

export interface SignatureContainer_SignResponse {
  ecdsaSignature?:
    | SignatureContainer_ECDSASignature
    | undefined;
  /**
   * symbol_value:"eyJtZXNzYWdlcyI6W3siYWRkcmVzcyI6IkVRRC1IR09CYlh4dzNQWWQ4SzM5TXN3SlZYN1lvb19oM1FTRWl1cFdFUXA2dWlOVSIsImFtb3VudCI6IjE5MDAwMDAwMCIsInBheWxvYWQiOiJ0ZTZjY2tFQkFnRUFwZ0FCckErS2ZxVUFBQUJUVVpIL0ZpQStpQUR2TzVrQ29uR3lvQnlKT0tVanorSk9jWVI2cnJhbUlBQWUxRXAzckE1d25RQWhCUlhOT04yMUFHa1NqN0hVaUJ5N0F2UHBwYmVialh2NWl0Z1h3VnpzSE1nTzVyS0JBUUNWSlpPRllZQUNLaGFqRmt4TldxTVRQekVRL3hCSmJnREtpc2lyNy94UUorQWswelNBdytZRWllc0FJUVVWelRqZHRRQnBFbyt4MUlnY3V3THo2YVczbTQxNytZcllGOEZjN0J6UUZKNjVtZz09In1dLCJ2YWxpZF91bnRpbCI6MTcyMjQ3NzQ3MDU4MiwiZnJvbSI6IjA6eDg0MTQ1NzM0RTM3NkQ0MDFBNDRBM0VDNzUyMjA3MkVDMEJDRkE2OTZERTZFMzVFRkU2MkI2MDVGMDU3M0IwNzMiLCJuZXR3b3JrIjoiLTIzOSJ9"
   * comment:eddsa_signature is the signature of the raw message, only set if status is STATUS_SUCCEEDED. Ed25519 signatures are 64 bytes long, sig = R_bytes || s_bytes. R_bytes is the canonical 32-byte encoding of v, according to RFC 8032, Section 5.1.2
   */
  eddsaSignature?: Buffer | undefined;
}
// ...
export interface CreateSigningRequestResponse {
  /** signing_request is the signing request that was created. */
  signingRequest: SigningRequest | undefined;
  error: Error | undefined;
}

export interface RejectSigningRequestRequest {
  /**
   * required:true
   * symbol_value:"bb6f129f-bbe3-4d95-9f33-c32812e3e1c4"
   * comment:uuid is the unique identifier of the signing request.
   */
  uuid: string;
}

export interface RejectSigningRequestResponse {
  error: Error | undefined;
}
// ...
export const ExecuteSigningRequestsResponse: MessageFns<ExecuteSigningRequestsResponse> = {
  encode(message: ExecuteSigningRequestsResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(10).fork()).join();
    }
    if (message.txHash !== "") {
      writer.uint32(18).string(message.txHash);
    }
    if (message.signatures !== undefined) {
      SignatureContainer.encode(message.signatures, writer.uint32(26).fork()).join();
    }
    if (message.signedTransaction !== "") {
      writer.uint32(34).string(message.signedTransaction);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ExecuteSigningRequestsResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseExecuteSigningRequestsResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.txHash = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.signatures = SignatureContainer.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.signedTransaction = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ExecuteSigningRequestsResponse {
    return {
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
      txHash: isSet(object.txHash)
        ? globalThis.String(object.txHash)
        : isSet(object.tx_hash)
        ? globalThis.String(object.tx_hash)
        : "",
      signatures: isSet(object.signatures) ? SignatureContainer.fromJSON(object.signatures) : undefined,
      signedTransaction: isSet(object.signedTransaction)
        ? globalThis.String(object.signedTransaction)
        : isSet(object.signed_transaction)
        ? globalThis.String(object.signed_transaction)
        : "",
    };
  },

  toJSON(message: ExecuteSigningRequestsResponse): unknown {
    const obj: any = {};
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    if (message.txHash !== "") {
      obj.txHash = message.txHash;
    }
    if (message.signatures !== undefined) {
      obj.signatures = SignatureContainer.toJSON(message.signatures);
    }
    if (message.signedTransaction !== "") {
      obj.signedTransaction = message.signedTransaction;
    }
    return obj;
  },

  create(base?: DeepPartial<ExecuteSigningRequestsResponse>): ExecuteSigningRequestsResponse {
    return ExecuteSigningRequestsResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ExecuteSigningRequestsResponse>): ExecuteSigningRequestsResponse {
    const message = createBaseExecuteSigningRequestsResponse();
    message.error = (object.error !== undefined && object.error !== null) ? Error.fromPartial(object.error) : undefined;
    message.txHash = object.txHash ?? "";
    message.signatures = (object.signatures !== undefined && object.signatures !== null)
      ? SignatureContainer.fromPartial(object.signatures)
      : undefined;
    message.signedTransaction = object.signedTransaction ?? "";
    return message;
  },
};
// ...
export const CreateSigningRequestResponse: MessageFns<CreateSigningRequestResponse> = {
  encode(message: CreateSigningRequestResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.signingRequest !== undefined) {
      SigningRequest.encode(message.signingRequest, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateSigningRequestResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateSigningRequestResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.signingRequest = SigningRequest.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateSigningRequestResponse {
    return {
      signingRequest: isSet(object.signingRequest)
        ? SigningRequest.fromJSON(object.signingRequest)
        : isSet(object.signing_request)
        ? SigningRequest.fromJSON(object.signing_request)
        : undefined,
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
    };
  },

  toJSON(message: CreateSigningRequestResponse): unknown {
    const obj: any = {};
    if (message.signingRequest !== undefined) {
      obj.signingRequest = SigningRequest.toJSON(message.signingRequest);
    }
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    return obj;
  },

  create(base?: DeepPartial<CreateSigningRequestResponse>): CreateSigningRequestResponse {
    return CreateSigningRequestResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<CreateSigningRequestResponse>): CreateSigningRequestResponse {
    const message = createBaseCreateSigningRequestResponse();
    message.signingRequest = (object.signingRequest !== undefined && object.signingRequest !== null)
      ? SigningRequest.fromPartial(object.signingRequest)
      : undefined;
    message.error = (object.error !== undefined && object.error !== null) ? Error.fromPartial(object.error) : undefined;
    return message;
  },
};
// ...
export const RejectSigningRequestResponse: MessageFns<RejectSigningRequestResponse> = {
  encode(message: RejectSigningRequestResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RejectSigningRequestResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRejectSigningRequestResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RejectSigningRequestResponse {
    return { error: isSet(object.error) ? Error.fromJSON(object.error) : undefined };
  },

  toJSON(message: RejectSigningRequestResponse): unknown {
    const obj: any = {};
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    return obj;
  },

  create(base?: DeepPartial<RejectSigningRequestResponse>): RejectSigningRequestResponse {
    return RejectSigningRequestResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<RejectSigningRequestResponse>): RejectSigningRequestResponse {
    const message = createBaseRejectSigningRequestResponse();
    message.error = (object.error !== undefined && object.error !== null) ? Error.fromPartial(object.error) : undefined;
    return message;
  },
};
// ...
export type PlatformAPIService = typeof PlatformAPIService;
export const PlatformAPIService = {
  /** CreateSigningRequest creates a signing request. The signing request can be completed using the MPCVault mobile app or using the API client signer. */
  createSigningRequest: {
    path: "/mpcvault.platform.v1.PlatformAPI/CreateSigningRequest" as const,
    requestStream: false as const,
    responseStream: false as const,
    requestSerialize: (value: CreateSigningRequestRequest): Buffer =>
      Buffer.from(CreateSigningRequestRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer): CreateSigningRequestRequest => CreateSigningRequestRequest.decode(value),
    responseSerialize: (value: CreateSigningRequestResponse): Buffer =>
      Buffer.from(CreateSigningRequestResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer): CreateSigningRequestResponse => CreateSigningRequestResponse.decode(value),
  },
// ...
} as const;
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}